password update and updated seeds

+++ schema.sql
@@ -18,7 +18,7 @@ CREATE TABLE department (
 CREATE TABLE `role` (
   id INT NOT NULL AUTO_INCREMENT,
   title varchar(30) NOT NULL,
-  salary DECIMAL (12,2) NOT NULL, 
+  salary INT (12) NOT NULL, 
   department_id INT NOT NULL,
   PRIMARY KEY (id)
 );
